thinkpad+waybar: start squeekboard in hidden stage, add toggle button
Because of the need to login from TTY, Sway will mostly be started
in laptop mode (as opposed to tablet/tent). It was inconvenient that
squeekboard popped up when typing on the keyboard.

Now, squeekboard only shows when turned to tablet mode (though currently
not supported by libinput) or when explicitly toggled from a new button
on the right side of Waybar.